pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ragprobe"
version = "0.1.0"
description = "Adversarial testing and evaluation framework for RAG pipelines and LLM agents"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
authors = [{ name = "Rohan Sagvekar" }]
keywords = ["rag", "llm", "evaluation", "testing", "adversarial", "prompt-injection"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "openai>=1.0",
    "anthropic>=0.20",
    "faiss-cpu>=1.7",
    "tiktoken>=0.6",
    "beautifulsoup4>=4.12",
    "requests>=2.31",
    "sec-edgar-downloader>=5.0",
    "numpy>=1.24",
    "rich>=13.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pydantic>=2.0.0",
]
openai = [
    "openai>=1.0",
]

[project.scripts]
ragprobe = "ragprobe.cli:main"

[project.urls]
Homepage = "https://github.com/RohanS1202/ragprobe"
Issues = "https://github.com/RohanS1202/ragprobe/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"

[tool.setuptools.packages.find]
where = ["."]
include = ["ragprobe*"]
